+++ src/views/BrowseInventory/BrowseInventory.js
@@ -1,3 +1,7 @@
+import { useCallback, useMemo } from 'react';
+import { useIntl } from 'react-intl';
+import { useHistory, useLocation } from 'react-router-dom';
+
 import {
   useNamespace,
 } from '@folio/stripes/core';
@@ -6,17 +10,74 @@ import {
 } from '@folio/stripes-smart-components';
 import {
   FiltersPane,
+  ResetButton,
+  SingleSearchForm,
   useFiltersToogle,
+  useLocationFilters,
 } from '@folio/stripes-acq-components';
 
 import {
+  BrowseInventoryFilters,
   BrowseResultsPane,
   SearchModeNavigation,
 } from '../../components';
+import { browseInstanceIndexes } from '../../filterConfig';
+import {
+  useBrowseValidation,
+  useInventoryBrowse,
+} from '../../hooks';
 
 const BrowseInventory = () => {
+  const history = useHistory();
+  const location = useLocation();
+  const intl = useIntl();
   const [namespace] = useNamespace();
   const { isFiltersOpened, toggleFilters } = useFiltersToogle(`${namespace}/filters`);
+
+  const [
+    filters,
+    searchQuery,
+    applyFilters,
+    applySearch,
+    changeSearch,
+    resetFilters,
+    changeSearchIndex,
+    searchIndex,
+  ] = useLocationFilters(location, history, () => {});
+
+  const {
+    data,
+    isFetching,
+    isLoading,
+    pagination,
+    totalRecords,
+  } = useInventoryBrowse({
+    filters,
+    searchIndex,
+  });
+
+  const { validateDataQuery } = useBrowseValidation(searchIndex);
+
+  const searchableIndexesPlaceholder = intl.formatMessage({ id: 'ui-inventory.browse.searchableIndexesPlaceholder' });
+  const isResetButtonDisabled = !location.search && !searchQuery;
+
+  const formattedSearchableIndexes = useMemo(() => (
+    browseInstanceIndexes.map(({ label, ...rest }) => ({
+      label: intl.formatMessage({ id: label }),
+      ...rest,
+    }))
+  ), []);
+
+  const onApplySearch = useCallback(() => {
+    const isSearchQUeryValid = validateDataQuery(searchQuery);
+
+    if (isSearchQUeryValid) applySearch();
+  }, [searchQuery]);
+
+  const onChangeSearchIndex = useCallback((e) => {
+    resetFilters();
+    changeSearchIndex(e);
+  }, []);
 
   return (
     <PersistedPaneset
@@ -29,13 +90,42 @@ const BrowseInventory = () => {
           toggleFilters={toggleFilters}
         >
           <SearchModeNavigation />
-          <>TODO: filters</>
+
+          <SingleSearchForm
+            applySearch={onApplySearch}
+            changeSearch={changeSearch}
+            disabled={!searchIndex}
+            searchQuery={searchQuery}
+            isLoading={isFetching}
+            ariaLabelId="ui-inventory.browse"
+            searchableIndexes={formattedSearchableIndexes}
+            changeSearchIndex={onChangeSearchIndex}
+            selectedIndex={searchIndex}
+            searchableIndexesPlaceholder={searchableIndexesPlaceholder}
+          />
+
+          <ResetButton
+            reset={resetFilters}
+            disabled={isResetButtonDisabled}
+          />
+
+          <BrowseInventoryFilters
+            activeFilters={filters}
+            applyFilters={applyFilters}
+            searchIndex={searchIndex}
+          />
         </FiltersPane>
       )}
 
       <BrowseResultsPane
+        browseData={data}
+        filters={filters}
+        isFetching={isFetching}
         isFiltersOpened={isFiltersOpened}
+        isLoading={isLoading}
+        pagination={pagination}
         toggleFiltersPane={toggleFilters}
+        totalRecords={totalRecords}
       />
     </PersistedPaneset>
   );